requirements.txt
Django
pillow
django-extensions
django-robots
django-debug-toolbar
django-taggit
django-summernote
django-multi-captcha-admin
django-simple-captcha
django-compressor
gunicorn
whitenoise
psycopg2-binary
